remove DataSource class
projects were designed to have more than 1 data source. But since this is not going to be implemented, to reduce the project complexity we assume that every project has only 1 data source. This makes the DataSource class useless. We moved the data_schema and the data frame to the project class and removed the DataSource class